Update app UI

+++ packages/server/src/services/challenge-svc.ts
@@ -4,22 +4,28 @@ import { Challenge } from "../models/index.ts";
 const challengeSchema = new Schema<Challenge>(
   {
     id: String,
+    owner: String,
     title: String,
     description: String,
     image: String,
-    link: String
+    link: String,
+    duration: String,
+    status: String,
+    stake: String,
+    participantOneGoal: String,
+    scoring: String
   },
   { collection: "challenges" }
 );
 
 const ChallengeModel = model<Challenge>("Challenge", challengeSchema);
 
-function index(): Promise<Challenge[]> {
-  return ChallengeModel.find().exec();
+function index(owner: string): Promise<Challenge[]> {
+  return ChallengeModel.find({ owner }).exec();
 }
 
-function get(id: string): Promise<Challenge | undefined> {
-  return ChallengeModel.findOne({ id })
+function get(id: string, owner: string): Promise<Challenge | undefined> {
+  return ChallengeModel.findOne({ id, owner })
     .exec()
     .then((challenge) => challenge ?? undefined);
 }
@@ -31,7 +37,7 @@ function create(json: Challenge): Promise<Challenge> {
 
 function update(id: string, challenge: Challenge): Promise<Challenge> {
   return ChallengeModel.findOneAndUpdate(
-    { id },
+    { id, owner: challenge.owner },
     challenge,
     { new: true }
   )
@@ -42,8 +48,8 @@ function update(id: string, challenge: Challenge): Promise<Challenge> {
     });
 }
 
-function remove(id: string): Promise<void> {
-  return ChallengeModel.findOneAndDelete({ id })
+function remove(id: string, owner: string): Promise<void> {
+  return ChallengeModel.findOneAndDelete({ id, owner })
     .exec()
     .then((deleted) => {
       if (!deleted) throw new Error(`${id} not deleted`);
